Move `ClickBehavior` types to its own file
