+++ frontend/src/metabase-types/api/index.ts
@@ -2,6 +2,7 @@ export * from "./activity";
 export * from "./automagic-dashboards";
 export * from "./bookmark";
 export * from "./card";
+export * from "./click-behavior";
 export * from "./collection";
 export * from "./dashboard";
 export * from "./data-app";
